rename pgp-key as Pubkey
