+++ extension/chrome/elements/compose-modules/compose-draft-module.ts
@@ -18,7 +18,7 @@ import { Xss } from '../../../js/common/platform/xss.js';
 import { ViewModule } from '../../../js/common/view-module.js';
 import { ComposeView } from '../compose.js';
 import { KeyStore } from '../../../js/common/platform/store/key-store.js';
-import { PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 
 export class ComposeDraftModule extends ViewModule<ComposeView> {
