+++ extension/chrome/elements/compose-modules/compose-my-pubkey-module.ts
@@ -4,9 +4,9 @@
 
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
 import { Catch } from '../../../js/common/platform/catch.js';
-import { KeyInfo } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { KeyInfo } from '../../../js/common/core/crypto/pubkey.js';
 import { Lang } from '../../../js/common/lang.js';
-import { PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { ViewModule } from '../../../js/common/view-module.js';
 import { ComposeView } from '../compose.js';
