+++ extension/chrome/elements/compose-modules/compose-send-btn-module.ts
@@ -9,8 +9,8 @@ import { Catch } from '../../../js/common/platform/catch.js';
 import { ComposeSendBtnPopoverModule } from './compose-send-btn-popover-module.js';
 import { GeneralMailFormatter } from './formatters/general-mail-formatter.js';
 import { GmailRes } from '../../../js/common/api/email-provider/gmail/gmail-parser.js';
-import { KeyInfo, Pubkey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
-import { PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { KeyInfo, Pubkey } from '../../../js/common/core/crypto/pubkey.js';
+import { PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 import { SendBtnTexts } from './compose-types.js';
 import { SendableMsg } from '../../../js/common/api/email-provider/sendable-msg.js';
 import { Str } from '../../../js/common/core/common.js';
