+++ extension/chrome/settings/index.ts
@@ -4,7 +4,7 @@
 
 import { Bm, BrowserMsg } from '../../js/common/browser/browser-msg.js';
 import { Ui } from '../../js/common/browser/ui.js';
-import { KeyInfo, PgpKey } from '../../js/common/core/crypto/pgp/pgp-key.js';
+import { KeyInfo, PgpKey } from '../../js/common/core/crypto/pubkey.js';
 import { Str, Url, UrlParams } from '../../js/common/core/common.js';
 import { ApiErr } from '../../js/common/api/error/api-error.js';
 import { Assert } from '../../js/common/assert.js';
