+++ extension/chrome/settings/modules/backup-automatic-module.ts
@@ -9,7 +9,7 @@ import { Settings } from '../../../js/common/settings.js';
 import { UnreportableError } from '../../../js/common/platform/catch.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
-import { PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 import { Assert } from '../../../js/common/assert.js';
 import { GoogleAuth } from '../../../js/common/api/google-auth.js';
 import { KeyStore } from '../../../js/common/platform/store/key-store.js';
