+++ extension/chrome/settings/modules/my_key.ts
@@ -8,8 +8,8 @@ import { Att } from '../../../js/common/core/att.js';
 import { Backend } from '../../../js/common/api/backend.js';
 import { Browser } from '../../../js/common/browser/browser.js';
 import { Buf } from '../../../js/common/core/buf.js';
-import { KeyInfo, Pubkey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
-import { PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { KeyInfo, Pubkey } from '../../../js/common/core/crypto/pubkey.js';
+import { PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { Url, Str } from '../../../js/common/core/common.js';
 import { View } from '../../../js/common/view.js';
