+++ extension/chrome/settings/modules/my_key_user_ids.ts
@@ -2,7 +2,7 @@
 
 'use strict';
 
-import { KeyInfo, PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { KeyInfo, PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 
 import { Assert } from '../../../js/common/assert.js';
 import { Url, Str } from '../../../js/common/core/common.js';
