+++ extension/chrome/settings/modules/security.ts
@@ -6,7 +6,7 @@ import { Backend, FcUuidAuth } from '../../../js/common/api/backend.js';
 import { ApiErr } from '../../../js/common/api/error/api-error.js';
 import { Assert } from '../../../js/common/assert.js';
 import { Catch } from '../../../js/common/platform/catch.js';
-import { KeyInfo } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { KeyInfo } from '../../../js/common/core/crypto/pubkey.js';
 import { Settings } from '../../../js/common/settings.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { Url } from '../../../js/common/core/common.js';
