+++ extension/chrome/settings/setup/setup-key-manager-autogen.ts
@@ -4,7 +4,7 @@
 
 import { SetupOptions, SetupView } from '../setup.js';
 
-import { PgpKey } from '../../../js/common/core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../../../js/common/core/crypto/pubkey.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { Url } from '../../../js/common/core/common.js';
 import { AcctStore } from '../../../js/common/platform/store/acct-store.js';
