+++ extension/js/common/api/api.ts
@@ -10,7 +10,7 @@ import { Att } from '../core/att.js';
 import { BrowserMsg } from '../browser/browser-msg.js';
 import { Buf } from '../core/buf.js';
 import { Catch } from '../platform/catch.js';
-import { Contact } from '../core/crypto/pgp/pgp-key.js';
+import { Contact } from '../core/crypto/pubkey.js';
 import { Dict } from '../core/common.js';
 import { Env } from '../browser/env.js';
 import { secureRandomBytes } from '../platform/util.js';
