+++ extension/js/common/api/email-provider/gmail/gmail.ts
@@ -13,14 +13,14 @@ import { Att } from '../../../core/att.js';
 import { BrowserMsg } from '../../../browser/browser-msg.js';
 import { Buf } from '../../../core/buf.js';
 import { Catch } from '../../../platform/catch.js';
-import { Contact } from '../../../core/crypto/pgp/pgp-key.js';
+import { Contact } from '../../../core/crypto/pubkey.js';
 import { Env } from '../../../browser/env.js';
 import { FormatError } from '../../../core/crypto/pgp/pgp-msg.js';
 import { Google } from '../../google.js';
 import { GoogleAuth } from '../../google-auth.js';
 import { Mime } from '../../../core/mime.js';
 import { PgpArmor } from '../../../core/crypto/pgp/pgp-armor.js';
-import { PgpKey } from '../../../core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../../../core/crypto/pubkey.js';
 import { SendableMsg } from '../sendable-msg.js';
 import { Xss } from '../../../platform/xss.js';
 import { KeyStore } from '../../../platform/store/key-store.js';
