+++ extension/js/common/api/wkd.ts
@@ -7,7 +7,7 @@ import { ApiErr } from './error/api-error.js';
 import { opgp } from '../core/crypto/pgp/openpgpjs-custom.js';
 import { Buf } from '../core/buf.js';
 import { Catch } from '../platform/catch.js';
-import { PgpKey } from '../core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../core/crypto/pubkey.js';
 import { PubkeySearchResult } from './pub-lookup.js';
 
 // tslint:disable:no-null-keyword
