+++ extension/js/common/browser/browser-msg.ts
@@ -11,10 +11,10 @@ import { BrowserMsgCommonHandlers } from './browser-msg-common-handlers.js';
 import { Buf } from '../core/buf.js';
 import { Catch } from '../platform/catch.js';
 import { Env } from './env.js';
-import { KeyDetails } from '../core/crypto/pgp/pgp-key.js';
+import { KeyDetails } from '../core/crypto/pubkey.js';
 import { PassphraseDialogType } from '../xss-safe-factory.js';
 import { PgpHash } from '../core/crypto/pgp/pgp-hash.js';
-import { PgpKey } from '../core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../core/crypto/pubkey.js';
 import { PgpMsg } from '../core/crypto/pgp/pgp-msg.js';
 import { Ui } from './ui.js';
 import { GlobalStoreDict, GlobalIndex } from '../platform/store/global-store.js';
