+++ extension/js/common/org-rules.ts
@@ -4,7 +4,7 @@
 
 import { Str } from './core/common.js';
 import { AcctStore } from './platform/store/acct-store.js';
-import { KeyAlgo } from './core/crypto/pgp/pgp-key.js';
+import { KeyAlgo } from './core/crypto/pubkey.js';
 
 type DomainRules$flag = 'NO_PRV_CREATE' | 'NO_PRV_BACKUP' | 'PRV_AUTOIMPORT_OR_AUTOGEN' | 'PASS_PHRASE_QUIET_AUTOGEN' |
   'ENFORCE_ATTESTER_SUBMIT' | 'NO_ATTESTER_SUBMIT' | 'NO_KEY_MANAGER_PUB_LOOKUP' | 'USE_LEGACY_ATTESTER_SUBMIT' |
