+++ extension/js/common/platform/store/passphrase-store.ts
@@ -5,7 +5,7 @@ import { AccountIndex, AcctStore, AcctStoreDict } from './acct-store.js';
 import { SessionStore } from './session-store.js';
 import { PromiseCancellation, Dict } from '../../core/common.js';
 import { Ui } from '../../browser/ui.js';
-import { PgpKey } from '../../core/crypto/pgp/pgp-key.js';
+import { PgpKey } from '../../core/crypto/pubkey.js';
 
 /**
  * Local or session store of pass phrases
